Update Makefile, add test workflow
The workflow comes from the mkiocccentry repo. This is an attempt to get
this working without the code populated (yet). The Makefile simply has
the rules added that seem to (me at this hour) be necessary to make this
work but it might be the Makefile has to be modified. It is not strictly
necessary to do this but it would be nice for later when (hopefully in
the next few months at most) the jparse code is actually moved here.