Attempt to restrict paths in CodeQL
This worked in the mkiocccentry repo and it is hoped that it works fine
here without modifications. If it fails it could be because of the fact
this is only a placeholder at this point but it hopefully would be easy
to fix. Later on when the parser is populated into the repo all the
tests will be run properly.