Remove references to jfmt, jval and jnamval
As these tools will be heavily modified and mostly rewritten (almost
entirely) and since they are incomplete but not being worked on for now
(as there are other higher priorities and this will be the case for some
time) they are being removed from the mkiocccentry repo
(https://github.com/ioccc-src/mkiocccentry)) for now. Thus these tools
should not be referenced as they will (in a day or two at most) not
exist in that repo. Later on when the time comes to develop them they
will likely be here (or else in the other repo and soon afterwards be
copied here) and thus they will be documented.