Update .gitignore
Ignore tarballs (compressed of any kind and otherwise), bzip2 compressed
files, compressed files (with Lempel-Ziv coding), gzip compressed files
and vim swap files (or some of them - it should not ever be a problem
but in case someone was doing something strange and there were many many
files open with vim at the time it should be ignored). This is done now
(though more will be added in time) because I currently have a tarball
in the directory as a backup (for certain definitions of 'backup').