Add to README.md note about populating the repo
The library and applications in this repo have some dependencies that
have to be worked into the repo. As the development of the parser was
done over time in another repo so too is the populating of this repo.

This means that others cannot yet easily use the code. The note in the
README.md file will be removed once this is resolved (hopefully not too
distant future but this depends on finding time).