help make jparse independent of dbg and dyn_array repos
We use `#if defined(MKIOCCCENTRY_SRC)` to determine of we
are in a case where the `dbg/` and `dyn_array/` trees are
under the same parent source tree, or if they are assumed
have their include files installed.

This requires one to update and re-install from the [dyn_array
repo](https://github.com/lcn2/dyn_array).