Update README.md about re-entrancy
The code is now in its own repo so it should not refer to when it is
done as it already is done.

The parser should be re-entrant (according to the way flex/bison work)
but as noted it has not been tested further so one should still use with
care. If a tool is developed that needs this functionality then it can
be tested at that time.