Improve how everything is built
To PREPARE FOR allowing this repo to be compiled INDEPENDENT of having
../dbg and ../dyn_array directories (with the appropriate libraries
built) the Makefiles now link in dbg.a and dyn_array via -ldbg and
-ldyn_array.  This change does depend on the renaming of the libraries
in the other repo to libdbg.a and libdyn_array.a which has been done in
a pull request. This MIGHT have to be changed some but this is not yet
clear.

Once the above is done it SHOULD be possible to get this to work on any
system as long as the dbg library and dyn_array library are both
installed. It is hoped that the workflow will also work at this point.

UNTIL THAT TIME, however, the WORKFLOWS WILL FAIL so THIS SHOULD BE
CONSIDERED A PREPARATION COMMIT. To test that everything is in order
(after the merge) the README.md file will be updated to discuss how to
build this repo. After that, assuming everything is okay, the only thing
that has to be done (here) is add the new tools that are being discussed
(a number of tools, at least four) that I and Landon Curt Noll will be
working on over the coming months (we both have some tasks that are of
higher priority).