rename MKIOCCCENTRY_SRC to INTERNAL_INCLUDE
We used:

```sh
sgit -e 's/MKIOCCCENTRY_SRC/INTERNAL_INCLUDE/g' *
```

to rename the `MKIOCCCENTRY_SRC` CPP symbol to
a more generic `INTERNAL_INCLUDE` CPP symbol.

This change is in sync with a recent change to
the dyn_array repo, and will be coordinated
with pending change to the mkiocccentry repo soon.