Fix more library name references
In Makefile and test_jparse/Makefile there were references to the
subdirectories for dbg and dyn_array; these have been removed.

There were some references to jparse.a still, not libjparse.a, and these
have been updated now too.

It is likely that the workflow here will still fail as some fixes had to
be made in the dyn_array repo that have not been merged yet. Once this
has happened I will make another commit to test this and fix accordingly
if necessary. When that is done issue #2 can be closed.