Add build script for CodeQL
I am unsure if this is sufficient as this is the first time I have used
it. The problem with the CodeQL without extra steps is that we require
the dbg and dyn_array facilities. As long as git clone, cd, make all and
sudo make install work (like with test.yml) then this should work. If it
does not then something else will have to be done.