Update README.md about dependencies
We require that both dbg and dyn_array are installed on the system. The
update now shows how to get them, compile them and install them.

This commit is also a test for the workflows of GitHub as it is hoped
that the other repos have had all the changes necessary for it to work
(it is already possible for users but it would be nice if GitHub
workflows also passed now; if not more will have to be done to solve
that).