Try fixing syntax error in codeql.yml
I seem to recall it possibly having a problem with 'analyse' instead of
'analyze' and that seems to be the only difference on the problem line.
The test.yml now works but this one showed up as a problem. If this does
not resolve it the syntax error more investigation will be needed.