improve the test_jparse dependency rules
We improve `make depend` in `test_jparse/Makefile` to prevent include
paths going up over the top of the tree and back down into it as some
systems (such as RHEL9 Linux) might do.

While this is only a minor change to the `make depend` in
`test_jparse/Makefile`, it prevents some systems from making
a `make depend` edit that is not needed.

Performed `make clobber depend all test` to test the above.