Add PREFIX to install/uninstall rules
The default PREFIX is /usr/local but for those who wish to have a
different location, say because of a system policy or for any other
reason, we now offer the PREFIX standard. To use one can do, for example
to install with the PREFIX /usr:

    make PREFIX=/usr

and if one wishes to uninstall (to deobfuscate their system of JSON :-) )
and they used a different PREFIX, say /usr, then they can do:

    make PREFIX=/usr uninstall

These will (respectively) install and remove the binaries to /usr/bin,
the library to /usr/lib and the header files to /usr/include/jparse.

The README.md has been updated to document this as well as various other
fixes (file links for example).

The README.md file refers to this functionality in the dbg and dyn_array
repos but this has not yet been added. It will be soon.