Add jparse_bug_report.sh
This script, which can be expanded if necessary, is a script that runs a
lot of tests, in order to help with reporting issues, and is based on
the bug_report.sh script that I wrote for the mkiocccentry repo back in
2022. In order to use it one can do:

    make bug_report

from the repo directory. There are two other rules that allow for some
useful (in certain situations) options but those are not documented (the
options are in the man page but the README.md does not discuss this as
it's most useful to not have these options when reporting a problem).

We do not anticipate a bug in the code but there could be situations
where a person can not compile or use the tools or library and so we
offer this script.

The script gives instructions including the log file to attach to the
reports.