Change 'IOCCC tools' to 'jparse tools'
The man pages used to say 'IOCCC tools' as they were for the IOCCC but
now that this is a separate repo this is no longer strictly true.

The date of the last update of the man pages have not been updated as it
takes more time and no documentation changes were made.