Typo fix in json_util.h
It is not even clear how this typo even happened but this part of the
header file is for code that was removed as these tools were unfinished
(due to illness) and since then the functionality is being reconsidered.
This part and some utility functions are very likely to change and quite
possibly will be rewritten entirely (and some removed).