enable virgin builds for the mkiocccentry repo
We added to `Makefile`:

```
LD_DIR=
LD_DIR2=
```

When `jparse/Makefile` invokes a rule in `jparse/test_jparse/Makefile`, it passes values as trailing args
to the `${MAKE} ... -C test_jparse .. LD_DIR2="${LD_DIR2}"`.  NOTE: Only `LD_DIR2` is passed.

When `jparse/Makefile` needs to link using `-ldbg` and/or `-ldyn_array`,
the `${LD_DIR}` value is used.  For example:

```Makefile
      ${CC} ${CFLAGS} $^ -lm -o $@ ${LD_DIR} -ldbg -ldyn_array
```

When `jparse/test_jparse/Makefile` needs to link using `-ldbg` and/or
`-ldyn_array`, the `${LD_DIRs}` value is used.  For example:

```Makefile
      ${CC} ${CFLAGS} $^ -lm -o $@ ${LD_DIR2} -ldyn_array -ldbg
```

FYI: When in [mkiocccentry repo](https://github.com/ioccc-src/mkiocccentry)
invokes rules in this repo's `Makefile`, it passes values such as:

```
LD_DIR= -L../dbg -L../dyn_array
LD_DIR2= -L../../dbg -L../../dyn_array
```

All this allows this repo to build independently, and it allows
the [mkiocccentry repo](https://github.com/ioccc-src/mkiocccentry)
to use a clone of this repo into its `jparse/` sub-directory
and to allow that repo to perform a "virgin build" where
`make uninstall` had been run for the dbg repo, the dyn_array repo
and for this repo.

NOTE: This repo, standalone, does **NOT** need to support such
a "virgin build".  That will **NOT** work here.  This commit is
designed to allow other repos, such as the
[mkiocccentry repo](https://github.com/ioccc-src/mkiocccentry)
to perform a "virgin build".