Add legacy_uninstall rule to Makefile
Although it is extremely unlikely that anyone used this repo yet it used
to install the header file (a single one, not multiple ones) to
/usr/local/include/jparse.h (instead of /usr/local/include/jparse/) and
the library was called jparse.a (in /usr/local/lib/jparse.a). The
legacy_uninstall rule now removes those for those very few who had it
(like I did).