Fix and enhance jparse_test.sh
The script referred to paths in the form of jparse/ as in the
mkiocccentry repo (where the parser originated) the parser is in the
jparse/ subdirectory. These have been removed so that for instance
jparse/jparse and jparse/test_jparse now are jparse and test_jparse/
respectively.

The -L option from jparse_test.sh has been removed as it is a problem
if location errors are not actually correct and since the make test rule
used -L anyway it is not a problem to remove it. The man page has always
explained how to update the files in case this is needed.

The script has a new option -f which forces the test on strings (in
those files) to have to fail rather than pass. It is obviously up to the
user to make sure they specify the right file with the right option but
this way one can check for invalid json strings too. The script does not
have a default check of the file as normally we want to validate json as
good (except for the -d option). This is something that might be good to
change but that can be decided later.

The script now only checks the json trees if -d is specified as it has
always been documented.

The test_jparse/Makefile test rule now runs the jparse_test.sh with -f
on a new file, json_teststr_fail.txt, but one may run the script
directly, specifying one or more files (including stdin via '-') just
like without -f.

These changes need to be tested in the mkiocccentry repo and in
particular the path changes and if necessary a pull request can be
opened.