Fix and improve jparse_test.sh for outside repos
The script now has a -Z topdir option to help the script locate the
directory where both the jparse tool and the test_jparse subdirectory
reside. If this option is not specified the script will try and locate
it by searching for the following paths:

    jparse.c
    ../jparse.c
    ../../jparse.c

stopping at the first one. The last one is a hack for the rare case
(like in the mkiocccentry repo) where there is a subdirectory jparse
with jparse/test_jparse, in another project (usually not recommended due
to complications but in some cases it is useful); the second is is for
when one is in the test_jparse subdirectory whereas otherwise it assumes
that one is in the top level directory. If the file cannot be found, for
instance one is in a deeper level subdirectory, then one must use the -Z
topdir option. We will not check any other locations and it is entirely
unsupported to try this from just any directory.

The fix here requires a slightly different invocation of the script. In
particular for the jparse tool one should not specify a directory for
the jparse tool (i.e. not ../jparse but ./jparse) and for files and
directories under test_jparse one must specify test_jparse before the
rest of the path as the script now assumes that one is in the top level
directory. Of course if one has a funny directory structure the script
could fail and what that really means is that this tool is supposed to
be run from a clone of the repo itself. To put it more simply, this
script assumes that one is in the top level directory of the repo even
if one is not, hence why the paths to tools and directories have
changed.

The test_jparse/Makefile rule rebuild_jparse_err_files has been moved to
the top level Makefile and the test_jparse/Makefile now will first cd to
the top level directory before running the rule, assuming that one is in
the subdirectory of course (as the top level Makefile won't use that
directory). This is necessary because as far as the script knows the
paths to the files and directories have changed and for the same reason
the error files have been rebuilt.

The jparse_test.sh script has been updated to version 1.0.5 04-09-2024.