Fix and add jstr_test.sh to make test
The script was broken due to files and paths that are not found here but
are found in mkiocccentry repo (where this originated) but since make
test did not run it it was not detected.

This commit shows more than ever that a new script is needed to run all
the different tests like the mkiocccentry repo does. This commit and the
previous commit do cause problems with make test in that repo and
workarounds/fixes will have to be introduced here to solve those
problems.