Add timestamps to most Makefile rules
Some rules had:

    ${S} echo "${OUR_NAME}: make $@ ending at: `date`"

and

    ${S} echo "${OUR_NAME}: make $@ starting at: `date`"

but others just had:

    ${S} echo "${OUR_NAME}: make $@ starting"

and

    ${S} echo "${OUR_NAME}: make $@ ending"

Those that had a starting/ending message now have a timestamp. Not all
should have this: for instance the all rule is silent if nothing is to
be done so it does not have any timestamp.

In rules in the top level Makefile that do something in another
directory (like test_jparse) but do not do anything in the top level
directory the 'nothing to do' message has been removed as it actually
did do something.