Change date format in jparse_test.sh
By habit and without thinking I put the date format in jparse_test.sh in
the international way rather than the way the rest of the version
strings are (the American Way (tm, maybe, but I don't know :-) ).
Although I would rather it the international way the other version
strings are fine as is and more than one date form (and more than one
form of English) is (are) used so to at least make them the same format
I have changed it back (not that it's any less confusing except when the
day is > 12).

Anyway, the version string is now: "1.0.5 04-09-2024" (the version
number itself does not need to be changed as no functionality changed).