Remove ambiguity in verge debug output
The text 'is in valid format' could be thought to be a typo for 'is
invalid format' so it now just says 'is valid format'. It is not the
most ideal wording but the ambiguity is at least gone: 'is a valid
format' doesn't seem to work either which is why it is just 'is valid
format'.