Add JPARSE_REPO_VERSION to jparse.h
This version string is for releases of the repo where if a commit is
made it can be updated, assuming it is worth doing so. This version is
not the same thing as the jparse tool or jparse library versions and are
not at the same level.

The CHANGES.md file has been updated (for the first time since this repo
was populated, in fact) to have this release along with a number of
recent changes that were made.