Improve jparse_test.sh
In the case one needs to not run the error location test, one can use
the new -L option.

In the case that no files are specified on the command line the script
now tries to read from the default JSON strings files, the good one and
the bad one (both even if -f is used).

New version of jparse_test.sh is 1.0.6 2024-09-06.