Remove/change IOCCC references where needed
The GitHub URL for issues in the man pages have been updated to this
repo's URL.

Where incorrect mkiocccentry.c was changed to jparse.c (and in one case,
jstr_test.8, other source files were mentioned as the script to be able
to read from the respective files).

In the test JSON files the mkiocccentry references were not removed as
they are valid as test files and that is where the parser was originally
developed.

In some other cases mkiocccentry was also correct to leave in.

Some typos were fixed as well.