Improve error messages if invalid JSON
The error reporting has been improved in the jparse(1) tool in the
following ways:

- If verbosity level is > 0, then it will show the invalid token and hex value
of that token (along with the line and column).
- If verbosity is not specified, then it will just show the syntax error (the
bad token with the line and column) and then the warning that the JSON tree is
NULL (just like if verbosity specified) and then the error message (from
jparse(1) itself).

The error files in the bad_loc have been updated as now the error output has
changed.

Remove the timestamps from the Makefiles. Although useful at times it
can be a bit too much text especially in subdirectories.

Updated jparse(1) version to 1.1.6 2024-09-07.