Fix rm in Makefiles' rules
The variable in front of the ${RM} should be ${Q}, not any of the
others.

Also the ${RM_V} is added to the ones that did not have it and those
that had it after another option had those other options moved to after
the ${RM_V}.

In the past the -v option was not in the ${RM} lines so the RM_V option
is now empty; if one wants to see what is removed they can do e.g.:

    make RM_V=-v clobber

The -r option to ${RM} is not used except for directories, as a safety
mechanism.