Fix #include paths in test_jparse
The paths in jnum_chk.h in particular were in the form of
'../../jparse/test_jparse/' which would assume that the clone of the
repo is in a directory called jparse. This is a relic from the past.

Add JSON parser version string to the test_jparse tools' -V option and
their -h option as well. Other tools in the repo have this change as
well but this is pending additional things that prevent a proper commit
(well at least for two but the others can come later as well).

Run make depend to update dependencies.