Rename Makefile variable Q_V_OPTION to VERBOSITY
Rename and add to many rules in the Makefiles the Q_V_OPTION to
VERBOSITY so one can more easily override the verbosity (as typing
VERBOSITY=1 is easier than typing Q_V_OPTION=3). For instance to run
make test with -v 3:

    make VERBOSITY=3 test

The default is still determined like before.

Many rules have had the -V ${VERBOSITY} added to the commands (or some
commands) and the GitHub test workflow file now uses VERBOSITY=3 to help
identify problems when there is a problem (at the possible expense of
having to sift through more but as problems almost always need more
details it is a useful thing to have).

Not all rules have the -v ${VERBOSITY} as not all rules should have it.
In many cases where this is true there is a comment in the Makefiles but
this is not perfect.