Correct CHANGES.md
An issue that was referred to has NOT been completely resolved yet but
was erroneously included in the CHANGES.md. This has been removed (in
particular jstrdecode -Q/-e options). Sorry for the confusion!

The details on this were removed from CHANGES.md but fortunately there
were other changes in that release.