Add a missing change to CHANGES.md
An important fix was made earlier in the Makefiles where tags related
rules referred to the dbg and dyn_array (but as a typo it was dyn_alloc)
directories (in the parent directory of this repo). This did not cause
any errors but it is still incorrect so it is worth noting.