Enhancements and bug fixes in jstr* tools
The jstrdecode(1) the -Q option did not work right in a number of ways.

The tool jstrdecode(1) has a new option -e which says to enclose each
decoded arg with escaped quotes. This option can be used with -Q.

In order to resolve these problems a new set of files were added,
jstr_util.h and jstr_util.c, which are strictly for the jstr* tools. A
new struct jstring was added which holds a buffer (that is allocated
already so no strdup() - due to the way the encoding/decoding functions
work and how they have to work) and a size of the buffer along with a
pointer to the next in the list. The new function alloc_jstr() allocates
a new struct jstring * adding an ALREADY allocated char * with its size
to it. This size is important due to the fact that these strings MAY
contain a NUL byte. The function free_jstring() is what it sounds like:
a function to free the memory of a single struct jstring *. With this
function too it is assumed that the char is allocated but since it MAY
contain NUL bytes it could leave a memory leak in some cases.

Making use of these functions and the struct, jstrencode(1) and
jstrdecode(1) now each have their own list (encoded strings and decoded
strings respectively) and a function each to add to the list and free
the list (this is done this way to simplify things). The function that
reads from a stream calls the function to add to the respective list, as
long as no errors occur (NULL pointers). On the other hand, one can (and
the tools do when not reading from stdin) call the alloc_jstr() function
directly and then the add to list function (if needed). The add function
adds to the end of the list so that the order is correct, and a list is
used because the char *s can hold NUL bytes which causes problems for
the dynamic array facility, and because for some options like jstrdecode
-Q we have to wait until after everything is processed, particularly due
to reading from stdin.