Add setlocale() call to jstr{en,de}code
This does not solve the UTF-8 decoding bug but it seems that it might be
required and in any case it should be done.

Removed the is_utf8() function from json_util.c as it appears it will
not be useful and it might even be correct (a copy has been made in case
this proves to be false).