Disable debug flags in Makefiles
Changed optimisation flags in the Makefiles to not specify -g3 as debug
symbols are almost useless when optimising and we have -O3 in use. During
debugging one can always use:

    make C_OPT="-g3" clobber all

to compile in debug symbols or have a file in the respective directories
(those needed) called makefile.local with the line:

    C_OPT= -g3

which is used for development purposes but should not normally be done.