Add function decode_json_string() for json_decode
To help simplify the json_decode() function (as it's quite long), this
new function takes the length and calculated decoded size as well as the
pointers (the block of memory, the return length and the has_nul as
well) and then allocates the 'char *', and assuming non NULL, does what
was the second half of the json_decode() function. As json_encode() is
much simpler it seems like at this time that something like this is not
needed. This new function is not static but it is entirely unclear if
that is necessary.

Add to struct json_string the bool unicode. Currently unused (just
initialised to false) the purpose will be to indicate whether or not the
string has any invalid unicode symbols found during decoding.