Add jparse.json
This file includes in JSON information about this repo and the primary
tools in it, as well as the URL of the tool and the authors. While this
file is not really necessary, some projects have something like this and
because this is a JSON parser it seemed like a nice idea. As new tools
are being thought of they might or might not be added to this new file
so there is a last updated member in the file ('timestamp').

Doing this I discovered some issues in comments in some C and header
files as well as the README.md file and these have been fixed as well.