Improve is_available.sh to check checknr(1)
The script test_jparse/is_available.sh now will do a trivial test on
checknr. Unlike the others, however, the test checks that the exit code
is NON-zero. This is because it checks that an erroneous man page is declared
invalid. The original version always returned 0 which meant that it was not as
easy to determine if it was in error (in the test suites - one had to run it
manually and look at the output). Thus to verify that it's reliable we check for
a non-zero status on a man page with an error in it.

The prep.sh script now skips the check_man rule if checknr cannot be found
or is unreliable.