Reject certain UTF-8 codepoints - issue #13
With this commit it might be that issue #13 is resolved. Codepoints that
were thought to be visible images might not be after all. The confusion,
if this is so, comes from the fact that in a web browser an image shows
up. But pasting this into other things, for instance in vim, and even
TextEdit (macOS), shows nothing but a blank space. This might make sense
as it is a control character.

Unless something comes up it seems that #13 is in good shape, except
that the json_utf8.[hc] should be cleaned up and so should the functions
json_decode() (if necessary) and decode_json_string() (if necessary).

The following good JSON files seem to be bad, as they have invalid UTF-8
bytes, and thus they have been moved to the bad/ subdirectory:

    test_jparse/test_JSON/./good/y_string_escaped_noncharacter.json
    test_jparse/test_JSON/./good/y_string_last_surrogates_1_and_2.json
    test_jparse/test_JSON/./good/y_string_unicode_U+1FFFE_nonchar.json
    test_jparse/test_JSON/./good/y_string_unicode_U+10FFFE_nonchar.json
    test_jparse/test_JSON/./good/y_string_unicode_U+FDD0_nonchar.json
    test_jparse/test_JSON/./good/y_string_unicode_U+FFFE_nonchar.json

.. with the y_ prefix changed to n_. This is because, in my
understanding, the JSON spec requires bad UTF-8 codepoints to be
rejected. If this is incorrect it will be easy to restore these files.

As they say, time will tell, but this seems to be good now.