Add more encode/decode tests to jstr_test.sh
The two new tests, respectively, test the output of jstrdecode using the
output of jstrencode as part of the command line; and jstrdecode using
the output of jstrencode using the output of jstrencode as part of the
command line. The second one compares the output to the file
test_jparse/jstr_test.txt.

The jstrencode and jstrdecode -t option still suggests that there should
be an encoding/decoding test but what those should be is not yet clear.