New release of all tools
Remove has_nul in struct json_string as UTF-8 should, it is my
understanding, not have a NUL byte.

Fix path in jsemcgen.sh. The repo we know that uses this might have a
problem here that has to be fixed but the path was broken for this repo
so it has to be done, and as the script has an option to change the tool
path, it should not be a problem as this is just the default path.

Update all tools and the release to be the same version after issue #13 was
resolved: "1.2.0 2024-10-09". "1.2.0" was chosen because it was the first one >
some of the versions and the others could be bumped up to it without any harm.