Add fix to UTF-8 code points in noncharacter range
The recommendation is that if a codepoint is in the so-called
non-character range, the code point should be set to the replacement
character (0xFFFD) which the code now does. As the UTF-8 length is 3
bytes an additional 3 bytes are added in the case a codepoint is in the
non-character range.

Set JPARSE_UTF8_VERSION to "1.2.2 2024-10-13".