Move bad test JSON file to good JSON directory
Due to commit 1a33062c6fb90bc2cc045354364e29fd23bd1794, which suggests
applications set unicode non-characters to the replacement character,
the file n_string_unicode_U+FDD0_nonchar.json had to be moved to the
good directory. It was renamed to
y_string_unicode_U+FDD0_replacement_char.json to suggest that a
replacement is made.

This is quite unfortunate as it makes our test suite diverge even more
from the test suite repo we found but according to
https://www.unicode.org/versions/Unicode13.0.0/UnicodeStandard-13.0.pdf
this should be done.