Rollback UTF-8 non-character fix
Upon further inspection it seems that a conforming program can either
reject it outright or change it to a designated replacement character.
As rejecting it seems more natural, and as it allows us to not diverge
as much from the JSON test suite we undo that modification.

The test JSON file was moved back to its original name under the bad
subdirectory.

The function is_unicode_noncharacter() still does exist, however, and
the macro UNICODE_REPLACEMENT_CHAR still does exist as well, though
currently it is unused.